requirements.txt
django
psycopg2==2.9.6
djangorestframework
django-filter
djangorestframework-simplejwt
